[SimpleWallet] Align outputs of "bc_height", "count_transfers" and "outputs" commands * [StringTools] Change first parameter of repeatChar() to size_t * [SimpleWallet] Add total number of transactions to "count_transactions"

+++ src/Common/StringTools.cpp
@@ -354,7 +354,7 @@ std::string timeIntervalToString(uint64_t intervalInSeconds) {
   return ss.str();
 }
 
-std::string repeatChar(int count, char c) {
+std::string repeatChar(size_t count, char c) {
   return std::string(count, c);
 }
